:recycle: [Refactor] 대여 서비스 공통 로직 분리
